% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fi_tsne.R
\name{fi_tsne}
\alias{fi_tsne}
\alias{fi_tsne.matrix}
\title{Run FI-tSNE algorithm}
\usage{
fi_tsne(x, ...)

\method{fi_tsne}{matrix}(
  x,
  n_components = 2L,
  n_jobs = 1L,
  learning_rate = "auto",
  perplexity = 30,
  early_exaggeration = 250,
  early_exaggeration_iter = 12L,
  n_iter = 500L,
  exaggeration = NULL,
  dof = 1,
  theta = 0.5,
  n_interpolation_points = 3L,
  min_num_intervals = 50,
  ints_in_interval = 1,
  initialization = c("pca", "spectral", "random"),
  metric = "euclidean",
  metric_params = NULL,
  initial_momentum = 0.5,
  final_momentum = 0.8,
  max_grad_norm = NULL,
  neighbors = c("auto", "exact", "annoy", "pynndescent", "approx"),
  negative_gradient_method = c("bh", "fft"),
  random_state = NULL,
  verbose = FALSE,
  ...
)
}
\value{
A matrix of t-SNE embeddings.
}
\description{
Run FI-tSNE algorithm
}
