% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fi_tsne.R
\name{embed}
\alias{embed}
\alias{embed.openTSNE.tsne.TSNEEmbedding}
\title{Embed new data in an existing t-SNE embedding object.}
\usage{
embed(x, ...)

\method{embed}{openTSNE.tsne.TSNEEmbedding}(x, new)
}
\arguments{
\item{x}{t-SNE embedding created with \code{\link{fi_tsne_embedding}}.}

\item{new}{New data to project into existing embedding}
}
\value{
Numeric matrix of t-SNE co-ordinates resulting from embedding
 \code{new} into the t-SNE embedding \code{x}.
}
\description{
Embed new data in an existing t-SNE embedding object.
}
\examples{
 set.seed(42)
 m <- matrix(rnorm(20000), ncol=20) 
 out_binding <- fi_tsne(m[-(1:2), ], as_matrix = FALSE, random_state = 42L)
 new_points <- embed(out_binding, m[1:2, ])
 plot(as.matrix(out_binding), col = "black", pch = 19,
     xlab = "t-SNE 1", ylab = "t-SNE 2")
 points(new_points, col = "red", pch = 19)
}
