% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fi_tsne.R
\name{embed}
\alias{embed}
\alias{embed.fi_tsne}
\title{Embed new data in an existing t-SNE embedding object.}
\usage{
embed(x, ...)

\method{embed}{fi_tsne}(x, new)
}
\arguments{
\item{x}{t-SNE embedding created with \code{\link{fi_tsne_embedding}}.}

\item{new}{New data to project into existing embedding}
}
\description{
Embed new data in an existing t-SNE embedding object.
}
